backend/requirements.txt
fastapi
uvicorn
pymupdf
langchain
langchain-text-splitters
sentence-transformers
faiss-cpu
requests
python-multipart
